requirements.txt
# ─── ElectON v2 — Production Dependencies ──────────────────────

# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
gunicorn>=22.0,<23.0
whitenoise>=6.6,<7.0

# Database
psycopg2>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Cache
django-redis>=5.4,<6.0
redis>=5.0,<6.0
hiredis>=3.0,<4.0

# Security
django-cors-headers>=4.3,<5.0
django-csp>=4.0,<5.0
PyJWT>=2.8,<3.0
cryptography>=42.0,<46.0

# Environment / Config
python-decouple>=3.8,<4.0

# Blockchain (Solana)
solana>=0.34,<1.0
anchorpy>=0.20,<1.0
solders>=0.21,<1.0

# File Processing
pandas>=2.1,<3.0
openpyxl>=3.1,<4.0
Pillow>=10.0,<11.0

# Cloud Storage (S3-compatible — Cloudflare R2 / DigitalOcean Spaces)
django-storages[s3]>=1.14,<2.0
boto3>=1.35,<2.0

# Reports
reportlab>=4.0,<5.0
matplotlib>=3.8,<4.0

# HTTP (webhooks + Brevo email API)
requests>=2.31,<3.0

# ─── Optional: Email Provider SDKs ────────────────────────────────
# Brevo backend uses the requests library above (no extra package needed).
#
# Azure Communication Services Email — activate by setting EMAIL_PROVIDER=azure
# or when BREVO_API_KEY is absent and AZURE_COMM_CONNECTION_STRING is set:
#   pip install azure-communication-email>=1.0,<2.0
# azure-communication-email>=1.0,<2.0

# Task Queue
celery>=5.3,<6.0
django-celery-beat>=2.6,<3.0

# Monitoring
sentry-sdk[django]>=2.0,<3.0

# Timezone
pytz>=2024.1
